+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - Dark mode.
 - [frontend] - Automatic title generation
 - [frontend] - Collapsible side bar.
+- [frontend] - Add confirmation box on thread deletion.
 
 ### Fixed
 - Migrate LLM Evaluation logic to scripts and add tests